pledge-user-salesforce/requirements.txt
google-cloud-error-reporting
simple-salesforce
google-cloud-secret-manager
sqlalchemy
pymysql